📦 NEW: Herobox initial implementation
