+++ src/App.tsx
@@ -10,7 +10,6 @@ import GameProvider from "./Components/GameProvider/GameProvider";
 import CreateOrJoinGame from "./Containers/CreateOrJoinGame/CreateOrJoinGame";
 import Game from "./Containers/Game/Game";
 import { Welcome } from "./Containers/Welcome";
-import WelcomePage from "./Containers/WelcomePage/WelcomePage";
 import { GameRoutePropsType, RoutePropsType } from "./Realms/Pure/types";
 
 interface Props {}
@@ -19,7 +18,6 @@ function App({}: Props): JSX.Element {
   return (
     <Router>
       <Switch>
-        <Route path="/welcome" exact component={WelcomePage} />
         <Route path="/" exact component={Welcome} />
         <Route path="/play/:gamecode" exact component={GamePage} />
         <Route
