+++ src/Containers/CreateOrJoinGame/CreateOrJoinGame.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { hot } from "react-hot-loader/root";
 import CreateGame from "../../Components/CreateGame/CreateGame";
 import JoinGame from "../../Components/JoinGame/JoinGame";
-import Layout from "../../Components/Layout/Layout";
+import Layout from "../../Components/Layout";
 import Nickname from "../../Components/Nickname/Nickname";
 import useJoinOrCreateGameContext from "../../Hooks/useCreateOrJoinGameContext/useCreateOrJoinGameContext";
